Move scripts and finish modpol.gp script.
The modpol.gp script can now compute the modular polynomials.
However, the method is naive (q-expansion of j) and inneficient.
A first improvement would be to determine the number of coefficients
of j needed to do the computation. Another would be to work in the
cyclotomic fields to avoid rounding. This might slow down the
computations even more.