Change theta_qexp to output in different formats (series, xs, list).
Also change theta_function to take as input the number of coeffs
to use to evaluate the q-expansion. The previous implementation
of theta_function didn't output the right result because of the
fact that the q-expansion of the theta series is sparse, so
suminf stops too early (see documentation on suminf for details).