Implement Zagier's closed formula to compute the fi.
Also implemented a first draft of a function to compute the class
polynomial using Borcherd's theorem (Theorem 3 in Zagier, Traces
of Singular Moduli). For the moment,this function works unless the class
polynomial is not really a polynomial. The other cases will be
implemented very soon.

Note that the closed formula of Zagier seems much slower that the
previously implemented recursive formula!